refactor(gateway): extract message/config helpers from run.py (slice 2 of #54962)
Second slice of the gateway god-file unpacking: extract the Slack
ignored-channel guard, CSV/list-to-set normalization, parent-channel
resolution, and message-timestamp opt-in check into
gateway/message_config_helpers.py with no behavior change.

- gateway/run.py: -65 lines; the 5 helpers are now imported from the
  module (module-attribute imports keep existing tests green)
- gateway/message_config_helpers.py: new module, pure functions, no
  module state — directly unit-testable

Follow-up to #77433 (display helpers, slice 1). Progress on #54962.
